FIRST_DELAY changed to long

+++ Project/src/main/java/com/tm/project/settings/Configuration.java
@@ -15,7 +15,7 @@ public class Configuration {
     public final boolean PROCESSING_COUNTERS = false;
     
     //ints
-    public final int FIRST_DELAY = 200;
+    public final long FIRST_DELAY = 200;
